\name{wa_ps_V811043_ts}
\docType{data}
\alias{wa_ps_V811043_ts}
\title{Waist Accelerometer data from a preschool aged subject V811043}
\description{
This data set gives counts/epoch(15 seconds) for subject in "Counts". The first column combines date and time information. 
}
\usage{wa_ps_V811043_ts}
\format{A matix 3 x 77157}
\source{Collected at the Energy Balance Laboratory at Vanderbilt University Medical Center (data owner), using a Actigraph GT1M accelerometer.}
\keyword{datasets}