\docType{data}
\name{wr_ad_lb1499}
\alias{wr_ad_lb1499}
\title{Wrist Accelerometer data from an adult subject lb1499}
\format{A dataset with 4,505 rows and 2 variables:
\describe{
\item{TS}{(TimeStamp) date and time in POSIXlt format.} 
\item{Axis 1}{counts/minute for the vertical axis.} 
}}
\source{Collected at the Energy Balance Laboratory at Vanderbilt University Medical Center (data owner), using a Actigraph GT1M accelerometer.}
\usage{data(wr_ad_lb1499)}
\description{
A dataset containing time information and counts/epoch.
}
\keyword{datasets}