\name{wr_ad_lb1499_ts}
\docType{data}
\alias{wr_ad_lb1499_ts}
\title{Wrist Accelerometer data from an adult subject lb1499}
\description{
This data set gives counts/minute for subject in "Axis 1". The first column is "TS" timestamp, and contains date and time information. 
}
\usage{wr_ad_lb1499_ts}
\format{A matix 2 x 4505}
\source{Collected at the Energy Balance Laboratory at Vanderbilt University Medical Center (data owner), using a Actigraph GT1M accelerometer.}
\keyword{datasets}