\name{zPhysActBedRest}
\docType{data}
\alias{zPhysActBedRest}
\title{Physical Activity Bed Rest Data}
\description{
This is the aforementioned csv's loaded as Rdata
}
%\usage{}
\format{matrices}
\source{Collected at the Energy Balance Laboratory at Vanderbilt University Medical Center (data owner), using a Actigraph GT1M accelerometer.}
\keyword{datasets}